backend/requirements.txt
langchain
python-dotenv
langchain_groq
langchain_google_genai
langchain_openai
langchain_community
pypdf
faiss-cpu
structlog
PyMuPDF
pandas
streamlit
langchain-text-splitters
langchain-core[tracing]
fastapi
uvicorn
pydantic
python-multipart
motor
pymongo
authlib
python-jose[cryptography]
scikit-learn
numpy
jwt
requests
pytest
pytest-asyncio
pytest-mock
httpx
tavily-python
twilio
firebase-admin
rank-bm25
youtube-transcript-api
beautifulsoup4
trafilatura
elevenlabs
openai>=1.0.0
